Refactored ProjectListing component.
This was a pretty big change in the structure of the ProjectListing
component by reducing the number of sub-components. Instead these were
converted to functions within the relevant components. I briefly
attempted to use do-expressions but the fact that this is still not
supported by default, I decided that it was better to get used to the
current methods of generating the output.